requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
openai
pytest
tiktoken
pandas